refactor(tasks): extract peer message broker
